refactor index.html and removed ui folders in entities & features
