+++ index.html
@@ -1,5 +1,5 @@
 <!DOCTYPE html>
-<html lang="en" dark-mode="dark">
+<html lang="en">
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
@@ -9,18 +9,17 @@
     <meta name="description" content="Management tool-shmool" />
     <meta property="og:title" content="Brello" />
     <meta property="og:description" content="Management tool-shmool" />
-
     <meta property="og:image" content="%VITE_SITE_URL%/social.jpg" />
     <meta property="og:image:alt" content="Brello" />
     <meta property="og:locale" content="en_GB" />
+    <meta property="og:locale:alternate" content="ru_RU" />
     <meta property="twitter:card" content="summary_large_image" />
     <meta property="theme-color" content="#FF00FF" />
+    <meta property="og:site_name" content="Brello Shmello" />
 
     <link rel="icon" href="/icon-%MODE%.svg" type="image/svg+xml" />
     <link rel="icon" href="/favicon.ico" size="any" />
-    <!-- 32x32  -->
     <link rel="apple-touch-icon" href="/apple-touch-icon.png" />
-    <!-- 180x180-->
     <link rel="manifest" href="/manifest.webmanifest" />
 
     <link rel="stylesheet" href="https://rsms.me/inter/inter.css" />
