--- src/widgets/header/model/index.ts
@@ -1,1 +1,0 @@
-export * from "./model";
